Select difficulty on lobby creation.

+++ Assets/Scripts/Game/CardItem/NormalCardItemController.cs
@@ -171,7 +171,7 @@ namespace Scripts
 
         private void SetColor(ProbabilityType probabilityType)
         {
-            _view.SetColor(ConstantValues.GetProbabilityTypeToColorMapping()[probabilityType]);
+            _view.SetColor(ConstantValues.GetProbabilityTypeToColorMapping()[(int)probabilityType]);
         }
 
         private void SetLockStatus(bool isLocked)
